fix: move product-context-display-properties services to export folders
