+++ src/app/extensions/punchout/exports/.gitignore
@@ -1,1 +1,1 @@
-*/**/*.*
+*/**/lazy*.*
