+++ src/app/extensions/tacton/exports/.gitignore
@@ -1,1 +1,1 @@
-*/**/*.*
+*/**/lazy*.*
