+++ src/app/extensions/tacton/exports/tacton-exports.module.ts
@@ -4,9 +4,8 @@ import { EXTERNAL_DISPLAY_PROPERTY_PROVIDER } from 'ish-core/facades/product-con
 import { FeatureToggleModule } from 'ish-core/feature-toggle.module';
 import { LAZY_FEATURE_MODULE } from 'ish-core/utils/module-loader/module-loader.service';
 
-import { TactonProductContextDisplayPropertiesService } from '../services/tacton-product-context-display-properties/tacton-product-context-display-properties.service';
-
 import { LazyTactonConfigureProductComponent } from './lazy-tacton-configure-product/lazy-tacton-configure-product.component';
+import { TactonProductContextDisplayPropertiesService } from './tacton-product-context-display-properties/tacton-product-context-display-properties.service';
 
 @NgModule({
   imports: [FeatureToggleModule],
